+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[UPDATE]**: Support previousValue in `TextField` format callback
 - [...]
 
 # v1.3.0 (21/03/2019)